microservice/telegram/requirements.txt
fastapi
uvicorn
telethon
python-multipart
pydantic-settings
sqlalchemy
aiosqlite
python-dotenv
